Able to save search OR queue as new playlist 👍
