+++ Spotiqueue/AppDelegate.swift
@@ -324,9 +324,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
                     && event.keyCode == kVK_Tab
                     && self.window.firstResponder == self.filterResultsField.currentEditor() {
             self.focusSearchBox()
-        } else if event.characters == "s"       // cmd-s = save current queue as playlist
-                    && flags == [.command] {
-            queueTableView.saveCurrentQueueAsPlaylist()
         } else {
             return event
         }
